More work on the annotation manager to make sure we can render properly. Still alpha at best.
Improved the repository view a bit to show WHY something was filtered.

Creating a new RepoDocDetail for how objects should be displayed.

Improved ProxyServerTest

Working on a new DocInfo advertisement system.

The 'added' member to DocInfo is not longer created when constructed since it's not added to the repository yet.

New PersistenceLayerTest to verify behavior.

Updated Broadcaster so I can broadcast manually.

Improved TestingTime so that we can time shift.

Migrated Http to typescript.