updated the main spectron test to open the PDF and PHZ files. They both work.
updates of more rules in tslint.

New AdvertisingPersistenceLayer so that each window can listen to persistence layer updates when files are deleted, created, and updated.

Datastore now supports delete of underlying files. Need to write tests for this though.

The Broadcast can now listen to an input channel and write to an output channel.