ok.. basic toggling of the sidebar and some cleanup.
We no longer show an empty sidebar with no elements and tell the user there
is nothing there.

More mockup of what the hover bar would look like when highlighting

Removed the bottom bar from the annotations sidebar for now.

Clean up of the MainAppMenu but not complete yet as I need to refactor it entirely so that we have reusable components.

Splitter reworked to be collapsed by default.