starting to work on a way to easily bulk import a large number of files - seems like it should be easy.
the reactor returns the event listener it registers so it can later be removed.

hopefully I fixed the problem with the annotation bar not reliably popping up. There's a second issue with not removing the annotation bar all the time but this isn't as much of an issue now.

spectron test to show the progress bar.